Update exporter.py
I have been using this library for marching cubes and export to file and noticed that the export was taking really long, so I change the function from using for loops to np.savetxt which is much faster and produces the same results. I benchmarked 1000 images and it was on average 2x faster!